bump copyright year
Note that I am not using a copyright year range here, on purpose.
As the template will be used when creating new projects, they will
generally only have copyright code starting the year of creation,
not dating back to the year of Software Heritage bootstrap. If
adaptation is needed, it should be done on a case-by-case basis.